Added an operational test on writing images

+++ src/render.rs
@@ -63,7 +63,7 @@ impl<T : Color> ImageData<T> {
             buffer.push(b);
         }
 
-        let _ = image::save_buffer(&Path::new(pathname), buffer.as_slice(), self.width, self.height, image::RGB(8)).unwrap();
+        let _ = image::save_buffer(&Path::new(pathname), buffer.as_slice(), self.width as u32, self.height as u32, image::RGB(8)).unwrap();
     }
 }
